#CSDH-84 test the like feature

+++ app/Http/Controllers/ReactionController.php
@@ -5,7 +5,11 @@ namespace App\Http\Controllers;
 use App\Models\Movie;
 use App\Models\Reaction;
 use App\Models\User;
+use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
+use InvalidArgumentException;
+use Throwable;
 
 class ReactionController extends Controller
 {
@@ -17,20 +21,21 @@ class ReactionController extends Controller
      */
     public function store(Request $request)
     {
-        $reactionData = $request->only(['reaction', 'user_id', 'movie_id']);
+        throw new Exception('Such reaction already exists');
+        try {
+            $reactionData = $request->only(['reaction', 'user_id', 'movie_id']);
 
-        // $user = User::where('id', $registerData['userId'])->first();
-        // $movie = Movie::where('id', $registerData['movieId'])->first();
+            $reaction = Reaction::where('user_id', $reactionData['user_id'])->where('movie_id', $reactionData['movie_id'])->first();
 
-        Reaction::where('user_id', $reactionData['user_id'])->where('movie_id', $reactionData['movie_id'])->delete();
-        $reaction = Reaction::create($reactionData);
+            if ($reaction['reaction'] === $reactionData['reaction'])
+                throw new InvalidArgumentException('Such reaction already exists');
 
-        // $reaction = new Reaction();
-        // $reaction->reaction = $registerData['reaction'];
-        // $reaction->user()->associate($user);
-        // $reaction->movie()->associate($movie);
-        // $reaction->save();
+            Reaction::where('user_id', $reactionData['user_id'])->where('movie_id', $reactionData['movie_id'])->delete();
+            $reaction = Reaction::create($reactionData);
 
-        return $reaction;
+            return $reaction;
+        } catch (Throwable $error) {
+            Log::debug($error);
+        }
     }
 }
